Fix 6 more compile errors: missing function names and precedence
- FUN_0602AAA0, FUN_0602AE24: Added missing function names to "long long" declarations
- FUN_060256CC, FUN_06025BF4, FUN_06026110, FUN_06027EDE: Fixed shift precedence

+++ src/FUN_06025BF4.c
@@ -92,13 +92,13 @@ void FUN_06025bf4()
 
         iVar10 = (uVar8 & 0xffff) << 1;
 
-        (*(int(*)())puVar2)(piVar11[1] << 2,*(int *)(0x06063750 + *((int)(int)piVar11 << 3)),
+        (*(int(*)())puVar2)(piVar11[1] << 2,*(int *)(0x06063750 + (*piVar11 << 3)),
 
                           ((unsigned int)(unsigned char)((char *)(0x06059074 + iVar10))[1] * 0x40 +
 
                           (unsigned int)(unsigned char)((int *)0x06059074)[iVar10]) << 1,
 
-                          0x00008000 + *(int *)((int)(0x06063750 + *((int)(int)piVar11 << 3)) + 4));
+                          0x00008000 + *(int *)((int)(0x06063750 + (*piVar11 << 3)) + 4));
 
         (*(int(*)())puVar2)(piVar11[1] << 2,
